requirements.txt
fastapi==0.128.0
openai==1.83.0
python-dotenv>=1.0.1,<2.0.0
pytest==8.3.4
uvicorn==0.40.0
